requirements.txt
EPCPyYes
quartet_epcis
quartet_masterdata
quartet_output
django-model-utils
factory_boy
